All review changes barring change to e.printStackTrace
