+++ nifi-extension-bundles/nifi-network-bundle/nifi-network-processors-nar/src/main/resources/META-INF/NOTICE
@@ -7,7 +7,9 @@ The Apache Software Foundation (http://www.apache.org/).
 This project includes derived works from the Kaitai Project available under the MIT License. 
   Portions of code found at https://formats.kaitai.io/pcap/java.html
   Copyright 2015-2023 Kaitai Project
-  The code can be found in nifi-nar-bundles/nifi-network-bundle/nifi-network-processors/PCAP.java
+  The code can be found in:
+    nifi-nar-bundles/nifi-network-bundle/nifi-network-processors/util/PCAP.java
+    nifi-nar-bundles/nifi-network-bundle/nifi-network-processors/util/Linktype.java
 ===========================================
 Apache Software License v2
 ===========================================
