+++ nifi-extension-bundles/nifi-network-bundle/nifi-network-processors/pom.xml
@@ -48,7 +48,6 @@
 		<dependency>
 			<groupId>org.junit.jupiter</groupId>
 			<artifactId>junit-jupiter-api</artifactId>
-			<version>5.11.0-M1</version>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
