+++ nifi-extension-bundles/nifi-network-bundle/nifi-network-processors/src/test/java/org/apache/nifi/processors/network/TestSplitPCAP.java
@@ -58,8 +58,7 @@ public class TestSplitPCAP {
                 0,  1,  2,  3,  4,  5,  6,  7,  8,  9,
                 10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
                 20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
-            },
-            "ETHERNET"
+            }
         );
 
         this.invalidPacket = new Packet(
@@ -71,8 +70,7 @@ public class TestSplitPCAP {
                 0,  1,  2,  3,  4,  5,  6,  7,  8,  9,
                 10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
                 20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
-            },
-            "ETHERNET"
+            }
         );
 
     }
